+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+## [2.0.0-nullsafety.1]
+
+* Cache functions with extra parameters.
+* Rename of cache functions to make them easier to use.
+
+Now, cache functions are:
+
+```dart
+cache1state((state) => () => ...);
+cache1state_1param((state) => (parameter) => ...);
+cache1state_2params((state) => (parameter1, parameter2) => ...);
+cache2states((state1, state2) => () => ...);
+cache2states_1param((state1, state2) => (parameter) => ...);
+cache2states_2params((state1, state2) => (parameter1, parameter2) => ...);
+cache3states((state1, state2, state3) => () => ...);
+cache1state_0params_x((state1, extra) => () => ...);
+cache2states_0params_x((state1, state2, extra) => () => ...);
+cache3states_0params_x((state1, state2, state3, extra) => () => ...);
+```    
+
 ## [1.4.0-nullsafety.0]
 
 * Migrating to null safety